remove extra state that is not used
The goal being to remove all the complexity that is not needed for the
working of the driver. Future improvement might not use all this state,
and previous state was growing quickly.

Proposal: make this driver a small merge-able basis upon which it is
possible to graft a lot of features like before.

Signed-off-by: Josuah Demangeon <josuah.demangeon@nordicsemi.no>